requirements.txt
django == 5.2.8
djangorestframework == 3.16.1
psycopg2-binary == 2.9.11
pytest == 9.0.0
pytest-django == 4.11.1
python-dotenv == 1.2.1
djangorestframework-simplejwt == 5.5.1
celery == 5.5.3
redis == 7.0.1
requests == 2.32.5
Faker == 37.12.0
